+++ apps/explorer/test/explorer/chain/csv_export/epoch_transactions_csv_exporter_test.exs
@@ -205,65 +205,65 @@ defmodule Explorer.CSV.Export.EpochTransactionsCsvExporterTest do
           }
         end)
 
-      assert result_voter.epoch_number == "902"
-      assert result_voter.block_number == to_string(block.number)
-      assert result_voter.timestamp == to_string(voter_reward.block_timestamp)
-      assert result_voter.epoch_tx_type == "Voter Rewards"
-      assert result_voter.validator_address == "N/A"
-      assert result_voter.validator_group_address == from_address_hash_voter |> normalize_address()
-      assert result_voter.to_address == to_address.hash |> normalize_address()
-      assert result_voter.tx_currency == "CELO"
-      assert result_voter.tx_currency_contract_address == celo_address |> normalize_address()
-      assert result_voter.type == "IN"
-      assert result_voter.locked_gold == "1000.000000000000000123"
-      assert result_voter.activated_gold == "1000.000000000000000001"
-      assert result_voter.value == to_string(voter_reward.amount |> Wei.to(:ether))
-      assert result_voter.value_wei == to_string(voter_reward.amount)
+      assert result_voter.epoch_number == [[], "902"]
+      assert result_voter.block_number == [[], to_string(block.number)]
+      assert result_voter.timestamp == [[], to_string(voter_reward.block_timestamp)]
+      assert result_voter.epoch_tx_type == [[], "Voter Rewards"]
+      assert result_voter.validator_address == [[], "N/A"]
+      assert result_voter.validator_group_address == [[], from_address_hash_voter |> normalize_address()]
+      assert result_voter.to_address == [[], to_address.hash |> normalize_address()]
+      assert result_voter.tx_currency == [[], "CELO"]
+      assert result_voter.tx_currency_contract_address == [[], celo_address |> normalize_address()]
+      assert result_voter.type == [[], "IN"]
+      assert result_voter.locked_gold == [[], "1000.000000000000000123"]
+      assert result_voter.activated_gold == [[], "1000.000000000000000001"]
+      assert result_voter.value == [[], to_string(voter_reward.amount |> Wei.to(:ether))]
+      assert result_voter.value_wei == [[], to_string(voter_reward.amount)]
 
-      assert result_voter_2.epoch_number == "903"
-      assert result_voter_2.block_number == to_string(block_2.number)
-      assert result_voter_2.timestamp == to_string(voter_reward_2.block_timestamp)
-      assert result_voter_2.epoch_tx_type == "Voter Rewards"
-      assert result_voter_2.validator_address == "N/A"
-      assert result_voter_2.validator_group_address == from_address_hash_voter |> normalize_address()
-      assert result_voter_2.to_address == to_address.hash |> normalize_address()
-      assert result_voter_2.tx_currency == "CELO"
-      assert result_voter_2.tx_currency_contract_address == celo_address |> normalize_address()
-      assert result_voter_2.type == "IN"
-      assert result_voter_2.locked_gold == "unknown"
-      assert result_voter_2.activated_gold == "unknown"
-      assert result_voter_2.value == to_string(voter_reward.amount |> Wei.to(:ether))
-      assert result_voter_2.value_wei == to_string(voter_reward.amount)
+      assert result_voter_2.epoch_number == [[], "903"]
+      assert result_voter_2.block_number == [[], to_string(block_2.number)]
+      assert result_voter_2.timestamp == [[], to_string(voter_reward_2.block_timestamp)]
+      assert result_voter_2.epoch_tx_type == [[], "Voter Rewards"]
+      assert result_voter_2.validator_address == [[], "N/A"]
+      assert result_voter_2.validator_group_address == [[], from_address_hash_voter |> normalize_address()]
+      assert result_voter_2.to_address == [[], to_address.hash |> normalize_address()]
+      assert result_voter_2.tx_currency == [[], "CELO"]
+      assert result_voter_2.tx_currency_contract_address == [[], celo_address |> normalize_address()]
+      assert result_voter_2.type == [[], "IN"]
+      assert result_voter_2.locked_gold == [[], "unknown"]
+      assert result_voter_2.activated_gold == [[], "unknown"]
+      assert result_voter_2.value == [[], to_string(voter_reward.amount |> Wei.to(:ether))]
+      assert result_voter_2.value_wei == [[], to_string(voter_reward.amount)]
 
-      assert result_validator.epoch_number == "903"
-      assert result_validator.block_number == to_string(block_2.number)
-      assert result_validator.timestamp == to_string(validator_reward.block_timestamp)
-      assert result_validator.epoch_tx_type == "Validator Rewards"
-      assert result_validator.validator_address == "N/A"
-      assert result_validator.validator_group_address == from_address_hash_validator |> normalize_address()
-      assert result_validator.to_address == to_address.hash |> normalize_address()
-      assert result_validator.tx_currency == "cUSD"
-      assert result_validator.tx_currency_contract_address == cusd_address |> normalize_address()
-      assert result_validator.type == "IN"
-      assert result_validator.locked_gold == "N/A"
-      assert result_validator.activated_gold == "N/A"
-      assert result_validator.value == to_string(validator_reward.amount |> Wei.to(:ether))
-      assert result_validator.value_wei == to_string(validator_reward.amount)
+      assert result_validator.epoch_number == [[], "903"]
+      assert result_validator.block_number == [[], to_string(block_2.number)]
+      assert result_validator.timestamp == [[], to_string(validator_reward.block_timestamp)]
+      assert result_validator.epoch_tx_type == [[], "Validator Rewards"]
+      assert result_validator.validator_address == [[], "N/A"]
+      assert result_validator.validator_group_address == [[], from_address_hash_validator |> normalize_address()]
+      assert result_validator.to_address == [[], to_address.hash |> normalize_address()]
+      assert result_validator.tx_currency == [[], "cUSD"]
+      assert result_validator.tx_currency_contract_address == [[], cusd_address |> normalize_address()]
+      assert result_validator.type == [[], "IN"]
+      assert result_validator.locked_gold == [[], "N/A"]
+      assert result_validator.activated_gold == [[], "N/A"]
+      assert result_validator.value == [[], to_string(validator_reward.amount |> Wei.to(:ether))]
+      assert result_validator.value_wei == [[], to_string(validator_reward.amount)]
 
-      assert result_group.epoch_number == "902"
-      assert result_group.block_number == to_string(block.number)
-      assert result_group.timestamp == to_string(group_reward.block_timestamp)
-      assert result_group.epoch_tx_type == "Validator Group Rewards"
-      assert result_group.validator_address == from_address_hash_group |> normalize_address()
-      assert result_group.validator_group_address == "N/A"
-      assert result_group.to_address == to_address.hash |> normalize_address()
-      assert result_group.tx_currency == "cUSD"
-      assert result_group.tx_currency_contract_address == cusd_address |> normalize_address()
-      assert result_group.type == "IN"
-      assert result_validator.locked_gold == "N/A"
-      assert result_validator.activated_gold == "N/A"
-      assert result_group.value == to_string(group_reward.amount |> Wei.to(:ether))
-      assert result_group.value_wei == to_string(group_reward.amount)
+      assert result_group.epoch_number == [[], "902"]
+      assert result_group.block_number == [[], to_string(block.number)]
+      assert result_group.timestamp == [[], to_string(group_reward.block_timestamp)]
+      assert result_group.epoch_tx_type == [[], "Validator Group Rewards"]
+      assert result_group.validator_address == [[], from_address_hash_group |> normalize_address()]
+      assert result_group.validator_group_address == [[], "N/A"]
+      assert result_group.to_address == [[], to_address.hash |> normalize_address()]
+      assert result_group.tx_currency == [[], "cUSD"]
+      assert result_group.tx_currency_contract_address == [[], cusd_address |> normalize_address()]
+      assert result_group.type == [[], "IN"]
+      assert result_validator.locked_gold == [[], "N/A"]
+      assert result_validator.activated_gold == [[], "N/A"]
+      assert result_group.value == [[], to_string(group_reward.amount |> Wei.to(:ether))]
+      assert result_group.value_wei == [[], to_string(group_reward.amount)]
     end
   end